chore: move account related pages into Account folder
